requirements.txt
 # Requirements for the PosterMaker application
pillow
pyside6
pytest